[REF] refactor wizard in three step files
