+++ account_invoice_invoice2data/wizards/__init__.py
@@ -1,2 +1,5 @@
 from . import wizard_invoice2data_import
+from . import wizard_invoice2data_import_state_import
+from . import wizard_invoice2data_import_state_map
+from . import wizard_invoice2data_import_state_apply
 from . import wizard_invoice2data_import_line
