+++ account_invoice_invoice2data/wizards/wizard_invoice2data_import.py
@@ -6,15 +6,8 @@
 # @author: Alexis de Lattre <alexis.delattre@akretion.com>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 
-import base64
-import mimetypes
-import os
-import tempfile
 
-import invoice2data
-
-from odoo import _, fields, models, tools
-from odoo.exceptions import UserError
+from odoo import fields, models
 
 
 class WizardInvoice2dataImport(models.TransientModel):
@@ -99,133 +92,3 @@ class WizardInvoice2dataImport(models.TransientModel):
         self.state = state
         action["res_id"] = self.id
         return action
-
-    def import_invoice(self):
-        self.ensure_one()
-        result = self._extract_json_from_pdf()
-        self._initialize_wizard_invoice(result)
-        self._initialize_wizard_lines(result)
-        if not all(self.mapped("line_ids.is_product_mapped")):
-            return self._get_action_from_state("product_mapping")
-        else:
-            self._analyze_invoice_lines()
-            return self._get_action_from_state("line_differences")
-
-    def _initialize_wizard_invoice(self, result):
-        for invoice_field in [
-            "amount_untaxed",
-            "amount",
-            "invoice_number",
-            "date",
-            "date_due",
-        ]:
-            if invoice_field in result:
-                value = result[invoice_field]
-                if "date" in invoice_field:
-                    value = value.date()
-                setattr(self, "pdf_%s" % invoice_field, value)
-
-    def map_products(self):
-        self.line_ids._create_supplierinfo()
-        if not all(self.mapped("line_ids.is_product_mapped")):
-            return self._get_action_from_state("product_mapping")
-        else:
-            self._analyze_invoice_lines()
-            return self._get_action_from_state("line_differences")
-
-    def _analyze_invoice_lines(self):
-        self.line_ids._analyze_invoice_lines()
-        self.to_delete_invoice_line_ids = self.mapped(
-            "invoice_id.invoice_line_ids"
-        ).filtered(lambda x: x.id not in self.mapped("line_ids.invoice_line_id").ids)
-
-    def _initialize_wizard_lines(self, pdf_data):
-        self.line_ids.unlink()
-        WizardLine = self.env["wizard.invoice2data.import.line"]
-        WizardLine.create(WizardLine._prepare_from_pdf_data(self, pdf_data))
-
-    def _extract_json_from_pdf(self):
-        self.ensure_one()
-
-        # Load Templates
-        local_templates_dir = tools.config.get("invoice2data_templates_dir", False)
-
-        if not local_templates_dir:
-            local_templates_dir = self.env.context.get(
-                "invoice2data_templates_dir", False
-            )
-
-        if not local_templates_dir:
-            raise UserError(
-                _("'invoice2data_templates_dir' not set in the odoo Config File")
-            )
-        if not os.path.isdir(local_templates_dir):
-            raise UserError(_("%s not available.") % str(local_templates_dir))
-        templates = invoice2data.extract.loader.read_templates(local_templates_dir)
-        if not len(templates):
-            raise UserError(_("No Template found to for bill invoices analyze."))
-
-        # Get data, and check filetype
-        file_data = base64.b64decode(self.invoice_file)
-        filetype = mimetypes.guess_type(self.invoice_filename)
-        if not filetype or filetype[0] != "application/pdf":
-            raise UserError(_("Unimplemented file type : '%s'") % str(filetype))
-
-        # Write data in a temporary file
-        fd, tmp_file_name = tempfile.mkstemp()
-        try:
-            os.write(fd, file_data)
-        finally:
-            os.close(fd)
-
-        try:
-            result = invoice2data.main.extract_data(tmp_file_name, templates=templates)
-        except Exception as e:
-            raise UserError(_("PDF Invoice parsing failed. Error message: %s") % e)
-        if not result:
-            raise UserError(_("This PDF invoice doesn't match a known templates"))
-
-        return result
-
-    def apply_changes(self):
-        self.ensure_one()
-        lines_vals = [x._prepare_invoice_line_vals() for x in self.line_ids]
-
-        sequence = len(lines_vals)
-        for line in self.to_delete_invoice_line_ids:
-            sequence += 1
-            line_vals = {
-                "sequence": sequence,
-                "price_unit": 0,
-                "name": _(
-                    "%s\n"
-                    "[PDF analysis] Unit Price %s set to 0,"
-                    " because the line is not present in the PDF."
-                )
-                % (line.name, line.quantity),
-            }
-            lines_vals.append((1, line.id, line_vals))
-
-        vals = {
-            "invoice_line_ids": lines_vals,
-        }
-        if self.pdf_date:
-            vals.update(
-                {
-                    "date_invoice": self.pdf_date,
-                }
-            )
-        if self.pdf_date_due:
-            vals.update(
-                {
-                    "date_due": self.pdf_date_due,
-                }
-            )
-        if self.pdf_invoice_number:
-            vals.update(
-                {
-                    "reference": self.pdf_invoice_number,
-                }
-            )
-
-        self.invoice_id.write(vals)
